Release v1.6.0: Early Bird APC Injection & Thread Primitives
* Implemented the Early Bird APC injection technique via the snd_inj_apc_* engine for Shellcode, PE, and COFF payloads.
* Introduced the snd_thread_api_t primitive to support remote thread suspension, APC queueing, and resumption.
* Expanded the snd_process_api_t interface to include process creation capabilities for suspended targets.
* Consolidated scattered injection examples into two robust CLI tools: inject_classic and inject_apc.
* Fixed a critical unsigned integer underflow bug in snd_pe_rva_to_ptr preventing out-of-bounds memory accesses.

+++ src/parsers/pe/utils.c
@@ -56,7 +56,7 @@ PVOID snd_pe_rva_to_ptr(const snd_pe_parser_t *parser, DWORD rva, SIZE_T size) {
     DWORD section_alignment = SND_PE_GET_NT_FIELD(parser, OptionalHeader.SectionAlignment);
     DWORD file_alignment    = SND_PE_GET_NT_FIELD(parser, OptionalHeader.FileAlignment);
 
-    for (DWORD i = parser->sections_count - 1; i >= 0; i--) {
+    for (int i = (int)parser->sections_count - 1; i >= 0; i--) {
         PIMAGE_SECTION_HEADER section    = &parser->section_head[i];
         DWORD                 aligned_va = section->VirtualAddress;
         if (section_alignment >= SND_PAGE_SIZE) {
